A working tail has been implemented
Previously, the tail only worked on one axis. Moving left then up, for example, would mess the tail up (the axis was switched).
To fix this, the method in which the tail is calculated was changed. Before, we recorded the direction the player was going in, then do the reverse and draw a tail there. This time, we are recording the x and y position of the snake head, then moving backwards in the vector.